+++ apps/terasu-web/src/actions/companies/detail/philosophy-action/philosophy-info-add-action.ts
@@ -1,6 +1,6 @@
 "use server";
 import { prismaClient } from "@terasu/db";
-import type { companyDetailPhilosophyAddTypes } from "@terasu/schema/models/companyDetailPhilosophySchema";
+import type { CompanyDetailPhilosophyAddTypes } from "@terasu/schema/models/companyDetailPhilosophySchema";
 
 /**
  * 企業理念を登録するアクション
@@ -9,7 +9,7 @@ import type { companyDetailPhilosophyAddTypes } from "@terasu/schema/models/comp
  */
 export const philosophyInfoAddAction = async (
   id: string,
-  data: companyDetailPhilosophyAddTypes,
+  data: CompanyDetailPhilosophyAddTypes,
 ) => {
   try {
     await prismaClient.companyPhilosophy.create({
